+++ Qiita_SwiftUI/Presentation/Screen/SearchResult/Component/TagInformationViewModel.swift
